pyproject.toml

[project]
name = "estetica"
version = "0.1.0"
description = ""
authors = [
    {name = "mylenetomasi", email = "mylenetomasi.jaque@gmail.com"},
]
dependencies = [
    "django>=4.2.5",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
    "setuptools>=68.1.2",
    "drf-spectacular>=0.26.4",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.2.0",
    "whitenoise[brotli]>=6.5.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
]
[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"